requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.1.1
pydantic==1.10.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
apscheduler==3.10.1
# scikit-learn builds many C extensions and can fail to compile on some Windows/Python combos.
# Se precisar do IsolationForest, instale scikit-learn manualmente (conda recommended).
# scikit-learn==1.2.2
# pywebpush versão ajustada para evitar erro de distribuição visto no Windows
pywebpush==2.0.3
httpx==0.24.1
python-dotenv==1.0.0
pytest==7.4.0
bcrypt==4.0.1
# garantir versão compatível do pymongo com motor
pymongo==4.3.3

# validador de email necessário por pydantic[email]
email-validator==1.3.1
